+++ example/lib/main.dart
@@ -1,4 +1,3 @@
-import 'package:example/models/department_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_tab_sync/flutter_tab_sync.dart';
 
@@ -17,33 +16,52 @@ class MyApp extends StatelessWidget {
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
         useMaterial3: true,
       ),
-      home:  MyHomePage(),
+      home: MyHomePage(),
     );
   }
 }
 
 class MyHomePage extends StatelessWidget {
-  MyHomePage({super.key,});
+  MyHomePage({
+    super.key,
+  });
 
-  final List<DepartmentModel> departments=[
-    DepartmentModel(name: 'Department1', employees: [
-      'Wissam','Mohammad','Mikel','Ahmad'
-    ],),
-    DepartmentModel(name: 'Department2', employees: [
-      'Max','Robin','Kareem','Toni','Lubna'
-    ],),
-    DepartmentModel(name: 'Department3', employees: [
-      'Leen','Nour',
-    ],),
-    DepartmentModel(name: 'Department4', employees: [
-      'Jessica','Luka','Antonio',
-    ],),
-    DepartmentModel(name: 'Department5', employees: [
-      'Dani','Lora','Mariam'
-    ],),
-    DepartmentModel(name: 'Department6', employees: [
-      'Jessica','Luka','Antonio',
-    ],),
+  final List<DepartmentModel> departments = [
+    DepartmentModel(
+      name: 'Department1',
+      employees: ['Wissam', 'Mohammad', 'Mikel', 'Ahmad'],
+    ),
+    DepartmentModel(
+      name: 'Department2',
+      employees: ['Max', 'Robin', 'Kareem', 'Toni', 'Lubna'],
+    ),
+    DepartmentModel(
+      name: 'Department3',
+      employees: [
+        'Leen',
+        'Nour',
+      ],
+    ),
+    DepartmentModel(
+      name: 'Department4',
+      employees: [
+        'Jessica',
+        'Luka',
+        'Antonio',
+      ],
+    ),
+    DepartmentModel(
+      name: 'Department5',
+      employees: ['Dani', 'Lora', 'Mariam'],
+    ),
+    DepartmentModel(
+      name: 'Department6',
+      employees: [
+        'Jessica',
+        'Luka',
+        'Antonio',
+      ],
+    ),
   ];
 
   @override
@@ -55,39 +73,53 @@ class MyHomePage extends StatelessWidget {
         surfaceTintColor: Colors.transparent,
         shadowColor: Colors.transparent,
         elevation: 0,
-
       ),
       body: LabeledTabViewSync<DepartmentModel>(
         items: departments,
-        itemBuilder: (department,isSelected)=>Column(
+        itemBuilder: (department, isSelected) => Column(
           crossAxisAlignment: CrossAxisAlignment.stretch,
           children: [
-            Text(
-                department.name,style: TextStyle(
-              fontSize: 20,
-              color: isSelected?Colors.white:Colors.black38,
-            )),
-            const SizedBox(height: 8,),
-            ...department.employees.map((employee)=>ListTile(
-              title: Text(
-                employee,style: TextStyle(color: isSelected?Colors.white:Colors.black38,),
-              ),
-            ))
+            Text(department.name,
+                style: TextStyle(
+                  fontSize: 20,
+                  color: isSelected ? Colors.white : Colors.black38,
+                )),
+            const SizedBox(
+              height: 8,
+            ),
+            ...department.employees.map((employee) => ListTile(
+                  title: Text(
+                    employee,
+                    style: TextStyle(
+                      color: isSelected ? Colors.white : Colors.black38,
+                    ),
+                  ),
+                ))
           ],
         ),
-        tabBuilder: (DepartmentModel item, bool isSelected)=>Text(
-          item.name,style: TextStyle(
-          fontSize: 12,
-          color: isSelected?Colors.white:Colors.black38,
+        tabBuilder: (DepartmentModel item, bool isSelected) => SizedBox(
+          height: 24,
+          child: Text(
+            item.name,
+            style: TextStyle(
+              fontSize: 12,
+              color: isSelected ? Colors.white : Colors.black38,
+            ),
+          ),
         ),
-        ),
-
         labelStyle: LabelStyle(
           color: Colors.deepPurple,
-          height: 32,
-          borderRadius: BorderRadius.circular(4),
+          height: 24,
         ),
+        barStyle: const BarStyle(height: 32, padding: EdgeInsets.zero),
       ),
     );
   }
+}
+
+class DepartmentModel {
+  final String name;
+  final List<String> employees;
+
+  DepartmentModel({required this.name, required this.employees});
 }