+++ example/lib/models/department_model.dart
@@ -1,6 +1,0 @@
-class DepartmentModel {
-  final String name;
-  final List<String> employees;
-
-  DepartmentModel({required this.name, required this.employees});
-}