+++ lib/flutter_tab_sync.dart
@@ -1,3 +1,6 @@
-export 'src/widgets/labeled_tab_view_sync.dart';
 export 'src/widgets/indexed_list_sync.dart';
+export 'src/widgets/labeled_tab_view_sync.dart';
+export 'src/widgets/indicated_tab_view_sync.dart';
 export 'src/styles/label_style.dart';
+export 'src/styles/indicator_style.dart';
+export 'src/styles/bar_style.dart';