+++ lib/src/widgets/labeled_tab_view_sync.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_tab_sync/src/styles/bar_style.dart';
 import 'package:flutter_tab_sync/src/styles/label_style.dart';
-import 'package:flutter_tab_sync/src/widgets/indexed_list_sync.dart';
 import 'package:flutter_tab_sync/src/widgets/labeled_tab_bar_sync.dart';
+import 'package:flutter_tab_sync/src/widgets/tab_view_sync.dart';
 
-class LabeledTabViewSync<T> extends StatefulWidget {
+class LabeledTabViewSync<T> extends StatelessWidget {
   /// the items which will use to construct tabs and the body
   final List<T> items;
-  final double tabsSpacing;
 
   /// the space between each two tabs , also it will be added at the beginning of the tab bar
   final double? itemsSpacing;
@@ -15,20 +15,17 @@ class LabeledTabViewSync<T> extends StatefulWidget {
   /// the space between the tab bar and the body
   final double? spacer;
 
-  /// the height of the bar
-  final double barHeight;
-
   /// this function will generate a tab for each item
   final Widget Function(T item, bool isSelected) tabBuilder;
 
   /// this function will generate the items in the body
   final Widget Function(T item, bool isSelected) itemBuilder;
 
-  /// padding for the tab bar
-  final EdgeInsets? tabBarPadding;
-
   /// padding for the body
   final EdgeInsets? bodyPadding;
+
+  /// the style of the bar
+  final BarStyle? barStyle;
 
   /// the style for the label behind the selected tab
   final LabelStyle labelStyle;
@@ -42,75 +39,30 @@ class LabeledTabViewSync<T> extends StatefulWidget {
     required this.tabBuilder,
     required this.itemBuilder,
     this.bodyPadding,
-    this.tabBarPadding,
+    this.barStyle,
     required this.labelStyle,
-    this.tabsSpacing = 8,
-    this.barHeight = 32,
     this.customViewBuilder,
   });
 
   @override
-  State<LabeledTabViewSync<T>> createState() => _LabeledTabViewSyncState<T>();
-}
-
-class _LabeledTabViewSyncState<T> extends State<LabeledTabViewSync<T>> {
-  ValueNotifier<int> selectedTabIndex = ValueNotifier(0);
-  ValueNotifier<int> selectedItemIndex = ValueNotifier(0);
-
-
-  Future<void> changeSelectedTab(final int newIndex,final bool isFromList) async {
-
-    if(newIndex==selectedTabIndex.value) {
-      return ;
-    }
-    selectedTabIndex.value = newIndex;
-  }
-  Future<void> changeSelectedItem(final int newIndex,final bool isFromList) async {
-
-    if(newIndex==selectedItemIndex.value) {
-      return ;
-    }
-    selectedItemIndex.value = newIndex;
-    selectedTabIndex.value = newIndex;
-  }
-
-  @override
   Widget build(final BuildContext context) {
-    final Widget tabBar =  ValueListenableBuilder<int>(
-        valueListenable: selectedTabIndex,
-        builder: (final _, final value, final __) =>
-          LabeledTabBarSync(
-            items: widget.items,
-            onTabPressed:(i)=> changeSelectedItem(i,false),
-            selectedTabIndex: value,
-            tabBuilder: widget.tabBuilder,
-            labelStyle: widget.labelStyle,
-            padding: widget.tabBarPadding,
-            tabsSpacing: widget.tabsSpacing,
-            barHeight: widget.barHeight,
-          ));
-
-          final Widget tabView =  ValueListenableBuilder<int>(
-    valueListenable: selectedItemIndex,
-    builder: (final _, final value, final __) =>IndexedListSync<T>(
-            items: widget.items,
-            itemBuilder: widget.itemBuilder,
-            onScroll:(i)=> changeSelectedTab(i,true),
-            selectedIndex: value,
-          ));
-
-          return widget.customViewBuilder != null
-              ? widget.customViewBuilder!(tabBar, tabView)
-              : Column(
-                  crossAxisAlignment: CrossAxisAlignment.stretch,
-                  children: [
-                    tabBar,
-                    SizedBox(width: widget.spacer),
-                    Expanded(
-                      child: tabView,
-                    ),
-                  ],
-                );
+    return TabViewSync(
+      items: items,
+      tabBarBuilder: (selectedIndex, onTab) => LabeledTabBarSync(
+        items: items,
+        onTabPressed: onTab,
+        selectedTabIndex: selectedIndex,
+        tabBuilder: tabBuilder,
+        labelStyle: labelStyle,
+        padding: barStyle?.padding ?? const BarStyle().padding,
+        tabsSpacing: barStyle?.tabsSpacing ?? const BarStyle().tabsSpacing,
+        barHeight: barStyle?.height ?? const BarStyle().height,
+      ),
+      itemBuilder: itemBuilder,
+      spacer: spacer,
+      itemsSpacing: itemsSpacing,
+      bodyPadding: bodyPadding,
+      customViewBuilder: customViewBuilder,
+    );
   }
-
 }