+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+[Feat]: add indicated_tab_view_sync
+[Update]: enhance labeled_tab_view_sync and indexed_list_sync
+
 ## 0.0.1
 
-Initial version with LabeledTabSyncView and IndexedListSync.
+Initial version with LabeledTabViewSync and IndexedListSync.